requirements.txt
alabaster==0.7.9
Babel==2.3.4
coverage==4.2
Django==1.10
django-allauth==0.27.0
docutils==0.12
html==1.16
imagesize==0.7.1
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
nameparser==0.5.1
numpy==1.11.1
oauthlib==2.0.0
pandas==0.18.1
psycopg2==2.6.2
Pygments==2.1.3
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.2
selenium==2.53.6
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.6
xmltodict==0.10.2
